+++ src/components/editable-field/editable-field-interface.tsx
@@ -1,3 +1,5 @@
+import { WcsSize } from '../../shared-types';
+
 export type ValidateFn<T> = (value: T) => boolean;
 export type FormatFn<T> = (value: T) => string;
 
@@ -9,6 +11,15 @@ export interface EditableComponentUpdateEvent {
      */
     successHandler: () => void;
     errorhandler: () => void;
+}
+
+export const WcsEditableFieldSizeValues = ['m', 'l'] as const; // as const keyword is used to infer and preserve the exact literal values of an array or object.
+
+export type WcsEditableFieldSize = Extract<WcsSize, typeof WcsEditableFieldSizeValues[number]>;
+
+export function isWcsEditableFieldSize(size: string): size is WcsEditableFieldSize {
+    // @ts-ignore : ignore size type, as it is checked with WcsEditableFieldSizeValues
+    return WcsEditableFieldSizeValues.includes(size);
 }
 
 export type EditableFieldType = 'input' | 'textarea' | 'select';