+++ stories/components/editable-field/editable-field.stories.ts
@@ -1,12 +1,18 @@
 import { Meta, StoryFn } from '@storybook/web-components';
 import { html } from 'lit-html';
-import { EditableFieldType, FormatFn, ValidateFn } from '../../../src/components/editable-field/editable-field-interface';
+import {
+    EditableFieldType,
+    FormatFn,
+    ValidateFn,
+    WcsEditableFieldSize
+} from '../../../src/components/editable-field/editable-field-interface';
 // @ts-ignore
 import editableFieldDocumentation from './editable-field-documentation.md'
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 // @ts-ignore
 import { withActions } from '@storybook/addon-actions/decorator';
 import { getComponentArgs } from '../../utils/args-generation';
+
 
 const meta: Meta = {
     title: 'Components/Editable-field',
@@ -28,7 +34,17 @@ const meta: Meta = {
 };
 export default meta;
 
-const Template: StoryFn<Partial<{ errorMsg: string, formatFn: FormatFn<any>, label: string, readonly: boolean, type: EditableFieldType, validateFn: ValidateFn<any>, value: any }>> = (args) => html`
+const Template: StoryFn<Partial<
+    {
+        errorMsg: string,
+        formatFn: FormatFn<any>,
+        label: string,
+        readonly: boolean,
+        type: EditableFieldType,
+        validateFn: ValidateFn<any>,
+        size: WcsEditableFieldSize,
+        value: any
+    }>> = (args) => html`
     <wcs-editable-field error-msg=${args.errorMsg}
                         .validateFn=${ifDefined(args.validateFn)}
                         .formatFn=${ifDefined(args.formatFn)}
@@ -36,21 +52,21 @@ const Template: StoryFn<Partial<{ errorMsg: string, formatFn: FormatFn<any>, lab
                         ?readonly=${args.readonly}
                         type=${args.type}
                         value=${args.value}
+                        size=${args.size}
                         id="editable-field-ex-1">
         ${args.type === 'input' ? html`
-            <wcs-input id="test"></wcs-input>
+            <wcs-input id="test" size=${args.size}></wcs-input>
         ` : ''}
         ${args.type === 'textarea' ? html`
             <wcs-textarea></wcs-textarea>
         ` : ''}
         ${args.type === 'select' ? html`
-            <wcs-select placeholder="Le select">
+            <wcs-select placeholder="Le select" size=${args.size}>
                 <wcs-select-option value="1">One</wcs-select-option>
                 <wcs-select-option value="2">Two</wcs-select-option>
                 <wcs-select-option value="3">Three</wcs-select-option>
             </wcs-select>
         ` : ''}
-    </wcs-editable-field>
 `;
 
 export const Default = Template.bind({});